improve HA
