+++ pkg/discordha/cache.go
@@ -1,0 +1,46 @@
+package discordha
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"time"
+
+	"go.etcd.io/etcd/clientv3"
+)
+
+var ErrorCacheKeyNotExist = errors.New("Cache key does not exist")
+
+func (h *HA) CacheRead(cache, key string, want interface{}) (interface{}, error) {
+	resp, err := h.etcd.Get(context.TODO(), fmt.Sprintf("/cache/%s/%s", cache, key))
+	if err != nil {
+		return nil, err
+	}
+
+	if resp.Count < 1 {
+		return nil, ErrorCacheKeyNotExist
+	}
+
+	err = json.Unmarshal(resp.Kvs[0].Value, &want)
+	if err != nil {
+		return nil, err
+	}
+
+	return want, nil
+}
+
+func (h *HA) CacheWrite(cache, key string, data interface{}, ttl time.Duration) error {
+	grant, err := h.etcd.Grant(context.TODO(), int64(ttl.Seconds()))
+	if err != nil {
+		return err
+	}
+
+	jsonData, err := json.Marshal(data)
+	if err != nil {
+		return err
+	}
+
+	_, err = h.etcd.Put(context.TODO(), fmt.Sprintf("/cache/%s/%s", cache, key), string(jsonData), clientv3.WithLease(grant.ID))
+	return err
+}
